requirements.txt
flask
dnspython
gunicorn